Move to temp class to show diff to Lucene
